+++ app/src/main/java/co/kr/woowahan_repo/application/AppContainer.kt
@@ -2,7 +2,7 @@ package co.kr.woowahan_repo.application
 
 import co.kr.woowahan_repo.data.api.ServiceCreator
 import co.kr.woowahan_repo.data.service.GithubIssuesService
-import co.kr.woowahan_repo.data.model.GithubIssuesRepositoryImpl
+import co.kr.woowahan_repo.data.model.repositoryimpl.GithubIssuesRepositoryImpl
 
 class AppContainer {
