+++ app/src/main/java/co/kr/woowahan_repo/di/ServiceLocator.kt
@@ -2,10 +2,10 @@ package co.kr.woowahan_repo.di
 
 import co.kr.woowahan_repo.BuildConfig
 import co.kr.woowahan_repo.data.api.interceptor.AuthInterceptor
-import co.kr.woowahan_repo.data.model.GithubOAuthRepositoryImpl
+import co.kr.woowahan_repo.data.model.repositoryimpl.GithubOAuthRepositoryImpl
 import co.kr.woowahan_repo.data.service.GithubOAuthAccessTokenService
 import co.kr.woowahan_repo.data.service.NotificationsService
-import co.kr.woowahan_repo.domain.GithubOAuthRepository
+import co.kr.woowahan_repo.domain.repository.GithubOAuthRepository
 import okhttp3.OkHttpClient
 import okhttp3.logging.HttpLoggingInterceptor
 import retrofit2.Retrofit
